requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.4.0
scipy>=1.7.0
scikit-image>=0.18.0
flask>=2.0.0
requests>=2.26.0
moviepy>=1.0.3
segment-anything
transformers
git+https://github.com/openai/CLIP.git
